`PsqlDosStorage`: Close all sessions when storage closes
The recommended approach for controlling session lifetime and the
associated connections is to open sessions using a context manager such
that they are automatically closed at the end of the scope.

The storage class currently does not use this approach but creates
sessions without context manager and so they have to be closed manually.
The `close` method should explicitly close any existing sessions.

The SqlAlchemy API provides the `sqlalchemy.orm.close_all_sessions`
function to close all sessions, but this will close _all_ sessions and
not just those that are associated with the current storage. Therefore,
instead we have to manually iterate over the global cache of sessions,
which is kept in `sqlalchemy.orm.session._sessions` and close those
sessions that are associated with the same engine as the storage's.